Move compositor_data.rs to compositor_layer.rs.
The new name better reflects the majority of the code in the file.
