+++ components/compositing/lib.rs
@@ -43,8 +43,7 @@ pub use constellation::Constellation;
 
 pub mod compositor_task;
 
-mod compositor_data;
-
+mod compositor_layer;
 mod compositor;
 mod headless;
